Develop

+++ generate.sh
@@ -2,14 +2,14 @@
 
 . ./.common.sh
 
-
 echo "${bold}*************************************"
 echo "Generate node key and a genesis file"
 echo "*************************************${normal}"
 if [ ! -f "ibftConfigFile.json" ]; then
-    echo "file not found"
+    echo "ibftConfigFile not found"
     exit 1
 fi
+
 besu operator generate-blockchain-config --config-file=ibftConfigFile.json --to=networkFiles --private-key-file-name=key
 keys="$(ls networkFiles/keys)"
 keys=($keys)
@@ -25,6 +25,10 @@ echo "${bold}*************************************"
 echo "Copy the genesis file to the network directory"
 echo "*************************************${normal}"
 for i in `seq 0 3`; do
-    cp networkFiles/keys/${keys[i]}/* node${i}/data
+    if [ ! -e node${i}/data/key ] && [ ! -e node${i}/data/key.pub ]; then
+        cp networkFiles/keys/${keys[i]}/* node${i}/data
+    else
+        echo "node${i} already exists"
+    fi
 done
 echo "Success\n"
